+++ backend/celery_worker.py
@@ -2,13 +2,13 @@ import asyncio
 import io
 import os
 
-import sentry_sdk
-from celery import Celery
 from celery.schedules import crontab
+from celery_config import celery
 from fastapi import UploadFile
 from logger import get_logger
 from models.files import File
 from models.settings import get_supabase_client
+from modules.brain.integrations.Notion.Notion_connector import NotionConnector
 from modules.brain.service.brain_service import BrainService
 from modules.notification.dto.inputs import NotificationUpdatableProperties
 from modules.notification.entity.notification import NotificationsStatusEnum
@@ -20,49 +20,9 @@ from packages.files.processors import filter_file
 
 logger = get_logger(__name__)
 
-sentry_dsn = os.getenv("SENTRY_DSN")
-if sentry_dsn:
-    sentry_sdk.init(
-        dsn=sentry_dsn,
-        sample_rate=0.1,
-        enable_tracing=True,
-    )
-
-CELERY_BROKER_URL = os.getenv("CELERY_BROKER_URL", "")
-CELERY_BROKER_QUEUE_NAME = os.getenv("CELERY_BROKER_QUEUE_NAME", "quivr")
-
 onboardingService = OnboardingService()
 notification_service = NotificationService()
 brain_service = BrainService()
-
-if CELERY_BROKER_URL.startswith("sqs"):
-    broker_transport_options = {
-        CELERY_BROKER_QUEUE_NAME: {
-            "my-q": {
-                "url": CELERY_BROKER_URL,
-            }
-        }
-    }
-    celery = Celery(
-        __name__,
-        broker=CELERY_BROKER_URL,
-        task_serializer="json",
-        task_concurrency=4,
-        worker_prefetch_multiplier=1,
-        broker_transport_options=broker_transport_options,
-    )
-    celery.conf.task_default_queue = CELERY_BROKER_QUEUE_NAME
-elif CELERY_BROKER_URL.startswith("redis"):
-    celery = Celery(
-        __name__,
-        broker=CELERY_BROKER_URL,
-        backend=CELERY_BROKER_URL,
-        task_concurrency=4,
-        worker_prefetch_multiplier=1,
-        task_serializer="json",
-    )
-else:
-    raise ValueError(f"Unsupported broker URL: {CELERY_BROKER_URL}")
 
 
 @celery.task(name="process_file_and_notify")
@@ -71,6 +31,7 @@ def process_file_and_notify(
     file_original_name: str,
     brain_id,
     notification_id=None,
+    integration=None,
 ):
     try:
         supabase_client = get_supabase_client()
@@ -201,3 +162,12 @@ celery.conf.beat_schedule = {
         "schedule": crontab(minute="0", hour="0"),
     },
 }
+
+
+@celery.task(name="NotionConnectorLoad")
+def process_integration_brain_created_initial_load(brain_id, user_id):
+    notion_connector = NotionConnector(brain_id=brain_id, user_id=user_id)
+
+    pages = notion_connector.compile_all_pages()
+
+    print("pages: ", len(pages))