+++ backend/packages/files/parsers/audio.py
@@ -9,7 +9,9 @@ from models import File, get_documents_vector_store
 from packages.files.file import compute_sha1_from_content
 
 
-async def process_audio(file: File, user, original_file_name):
+async def process_audio(
+    file: File, user, original_file_name, integration=None, integration_link=None
+):
     temp_filename = None
     file_sha = ""
     dateshort = time.strftime("%Y%m%d-%H%M%S")