+++ frontend/lib/types/BrainConfig.ts
@@ -41,6 +41,8 @@ export const openAiFreeModels = [
   "gpt-3.5-turbo",
   "gpt-3.5-turbo-1106",
   "gpt-3.5-turbo-16k",
+  "gpt-4-0125-preview",
+  "gpt-3.5-turbo-0125",
 ] as const;
 
 export const openAiPaidModels = [...openAiFreeModels, "gpt-4"] as const;