+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 0.0.195 (2024-02-06)
+
+## What's Changed
+* feat(integrations): integration with Notion in the backend by @StanGirard in https://github.com/StanGirard/quivr/pull/2123
+
+
+**Full Changelog**: https://github.com/StanGirard/quivr/compare/v0.0.194...v0.0.195
+
+## 0.0.194 (2024-02-05)
+
+## What's Changed
+* feat(frontend): add a chatbot for users  by @Zewed in https://github.com/StanGirard/quivr/pull/2144
+
+
+**Full Changelog**: https://github.com/StanGirard/quivr/compare/v0.0.193...v0.0.194
+
+## 0.0.193 (2024-02-04)
+
+## What's Changed
+* feat(frontend): design changes on user profile by @Zewed in https://github.com/StanGirard/quivr/pull/2140
+* fix(frontend): rename upgrade to plus to upgrade by @Zewed in https://github.com/StanGirard/quivr/pull/2141
+
+
+**Full Changelog**: https://github.com/StanGirard/quivr/compare/v0.0.192...v0.0.193
+
 ## 0.0.192 (2024-02-02)
 
 ## What's Changed